setting a new Location by calling ..../location/new/[newLocationName] view all Locations by calling .../location

+++ src/main/java/com/kum/daos/DeviceDAO.java
@@ -31,7 +31,7 @@ public class DeviceDAO {
 	}
 
 	public Device findByID(Device obj) throws SQLException {
-		if ((Integer)obj.getId() == null) {
+		if ((Long)obj.getId() == null) {
 			throw new SQLException("No ID!");
 		}
